Break up huge files and organize into directories
