+++ swift/Sources/stormlight-duel/game/Character/Paths.swift
@@ -1,0 +1,21 @@
+public struct PathProgress: Sendable {}
+
+public enum PathName: CaseIterable, Sendable {
+    // Heroic
+    case warrior
+    case agent
+    case envoy
+    case hunter
+    case leader
+    case scholar
+    // Radiant
+    case dustbringer
+    case edgedancer
+    case elsecaller
+    case lightweaver
+    case skybreaker
+    case stoneward
+    case truthwatcher
+    case willshaper
+    case windrunner
+}
