+++ swift/Sources/stormlight-duel/game/Combat/CombatAction/Move.swift
@@ -1,0 +1,12 @@
+public struct Move: CombatAction {
+    public var distanceToward: Distance
+    public static var actionCost: Int { 1 }
+    public func action(by character: RpgCharacterRef, in gameSession: isolated GameSession) async {
+        await gameSession.game.broadcaster.tellAll(
+            "\(character.name) moved \(distanceToward) toward their opponent")
+    }
+
+    public init(distanceToward: Distance) {
+        self.distanceToward = distanceToward
+    }
+}
