+++ swift/Sources/stormlight-duel/game/Condition/CompositeCondition/CompositeCondition.swift
@@ -1,0 +1,12 @@
+internal protocol CompositeCondition:
+    Condition,
+    NonLeafGenericListenerHolder,
+    AllTheListenersHolder
+{
+    associatedtype C: Condition
+    var core: C { get }
+}
+extension CompositeCondition {
+    public var id: Int { core.id }
+    public var childHolders: [Any] { [core] }
+}
