+++ swift/Sources/stormlight-duel/game/Condition/LeafCondition/LeafCondition.swift
@@ -1,0 +1,4 @@
+internal protocol LeafCondition: Condition, ListenerHolderLeaf, SelfListenerHolderLeaf,
+    SelfListenerSelfHookHolderLeaf, SelfListenerSelfHookForTestHolderLeaf
+{
+}
