+++ swift/Sources/stormlight-duel/game/Game/GameSnapshot.swift
