+++ swift/Sources/stormlight-duel/game/Item/Armor/Armor.swift
@@ -4,9 +4,3 @@ public protocol Armor: Item {
 }
 
 public protocol ArmorTrait {}
-
-public struct CumbersomeArmor: ArmorTrait {
-    var minStrength: Int
-}
-public struct Presentable: ArmorTrait {}
-public struct UniqueArmor: ArmorTrait {}
