+++ swift/Sources/stormlight-duel/game/Item/Armor/ArmorTraits/ArmorTraitStubs.swift
@@ -1,0 +1,5 @@
+public struct CumbersomeArmor: ArmorTrait {
+    var minStrength: Int
+}
+public struct Presentable: ArmorTrait {}
+public struct UniqueArmor: ArmorTrait {}
