+++ swift/Sources/stormlight-duel/game/Item/Item.swift
@@ -3,10 +3,19 @@ public protocol ItemSharedProtocol: Keyed where Key == ItemRef {
     var price: Money? { get }
     var weight: Weight { get }
 }
-
 extension ItemSharedProtocol {
     public var primaryKey: ItemRef {
         ItemRef(name: self.name)
+    }
+}
+
+public struct ItemRef: Hashable, Sendable {
+    public var name: String
+    init(name: String) {
+        self.name = name
+    }
+    public init(of item: any Item) {
+        self = item.primaryKey
     }
 }
 
@@ -14,9 +23,6 @@ public enum TraitCondition: Sendable {
     case always
     case expert
     case notExpert
-}
-
-public protocol ItemSnapshot: ItemSharedProtocol, Sendable {
 }
 
 public protocol Item: ItemSharedProtocol {
