+++ swift/Sources/stormlight-duel/game/Item/ItemSnapshot.swift
@@ -1,0 +1,2 @@
+public protocol ItemSnapshot: ItemSharedProtocol, Sendable {
+}
