+++ swift/Sources/stormlight-duel/game/Test/RpgTest.swift
@@ -17,9 +17,6 @@ extension RpgTestSharedProtocol {
     }
 }
 
-public protocol RpgTestSnapshot: RpgTestSharedProtocol, Sendable {
-}
-
 /// Since a test has lots of hooks, it is a fully mutable structure
 ///
 /// A head-to-head test is presented to each character as a separate test structure.
@@ -88,18 +85,6 @@ public class RpgSimpleTest: RpgTest {
     }
 }
 
-public struct RpgSimpleTestSnapshot: RpgTestSnapshot {
-    public let id: Int
-    public var skill: SkillName
-    public var difficulty: Int
-    public var advantages: Int
-    public var disadvantages: Int
-    public var testRolls: [RpgTestRoll]
-    public var opportunities: Int
-    public var complications: Int
-    public var success: Bool?
-}
-
 /// Use this in places where `any RpgTest` doesn't work.
 public class AnyRpgTest: RpgTest {
     public var core: any RpgTest
@@ -140,38 +125,6 @@ public class AnyRpgTest: RpgTest {
     }
     public var snapshot: any RpgTestSnapshot {
         core.snapshot
-    }
-}
-
-public struct AnyRpgTestSnapshot: RpgTestSnapshot {
-    public var core: any RpgTestSnapshot
-    public init(_ test: any RpgTestSnapshot) {
-        self.core = test
-    }
-    public var id: Int { core.id }
-    public var skill: SkillName {
-        core.skill
-    }
-    public var advantages: Int {
-        core.advantages
-    }
-    public var disadvantages: Int {
-        core.disadvantages
-    }
-    public var testRolls: [RpgTestRoll] {
-        core.testRolls
-    }
-    public var difficulty: Int {
-        core.difficulty
-    }
-    public var opportunities: Int {
-        core.opportunities
-    }
-    public var complications: Int {
-        core.complications
-    }
-    public var success: Bool? {
-        core.success
     }
 }
